move parsers and trackers to own resp. subdirs
