+++ tests/track/test_chunk_parser_plugin.py
@@ -1,8 +1,8 @@
 from datetime import timedelta
 
 from OTVision.dataformat import DETECTIONS
-from OTVision.track.chunk_parser_plugins import DetectionParser, JsonChunkParser
-from OTVision.track.Detection import Detection
+from OTVision.track.model.detection import Detection
+from OTVision.track.parser.chunk_parser_plugins import DetectionParser, JsonChunkParser
 from tests.track.test_data_builder import (
     DEFAULT_INPUT_FILE_PATH,
     DEFAULT_START_DATE,
